requirements.txt
gradio>=4.0.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
wikipedia>=1.4.0
wikipedia-api>=0.6.0
youtube-transcript-api>=0.6.1
yt-dlp>=2023.10.13
anthropic>=0.8.0
Pillow>=10.0.0
openpyxl>=3.1.2
transformers>=4.35.0
torch>=2.1.0
torchaudio>=2.1.0
sentencepiece>=0.1.99
litellm>=1.0.0
smolagents[litellm,transformers,all]>=0.2.1
mlx-whisper
protobuf>=3.20.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
matplotlib>=3.7.0